Fixed compiler warnings about comparison issues between 'int' and unsigned int (aka 'size_t')
In other words converted many loops in the form:

    for(int i = 0; i< schoolBuildings->size();i++){

to:

    for (size_t i = 0; i < schoolBuildings->size(); i++) {

The changes are:

1. Convert 'int' to 'size_t' which is the data type used by all C++
   data structures, because negative indexes or sizes (suggested by
   int) is not meangiful and only unsigned numbers should be
   used. Sadly, many languages do not support unsigned numbers and
   that forces programmers to have bad mental model and develop
   tererible code.

2. Spacing -- in English there should be 1 blank space between words.
   Similarly in C++ there should be a blank space between very
   word/operator as shown above.  Of course, these types of changes
   are automatically done by the editor (emacs/Eclipse/NetBeans etc.)
   but the programmer has to click 2 buttons for that happen.

   I have not reformatted the code fully to keep the deltas small to
   avoid breaking the code.

+++ haplos/src/Buildings/BuildingGenerator.cpp
@@ -191,7 +191,7 @@ void BuildingGenerator::generateBuildings(double businessSizeProbablities[6], do
         updateStatistics('S', capacity);
     }
     
-    for(int i = 0; i< schoolBuildings->size();i++){
+    for (size_t i = 0; i < schoolBuildings->size(); i++) {
         //Set up Pointers
         allBuildings->insert({schoolBuildings->at(i).getID(), &(schoolBuildings->at(i))});
         densityData->at(schoolBuildings->at(i).getLocation()[0]).at(schoolBuildings->at(i).getLocation()[1]).addBuilding(&(schoolBuildings->at(i)));
@@ -252,7 +252,7 @@ void BuildingGenerator::generateBuildings(double businessSizeProbablities[6], do
         updateStatistics('D', capacity);
 
     }
-    for(int i = 0; i< daycareBuildings->size();i++){
+    for (size_t i = 0; i < daycareBuildings->size(); i++) {
         //Set up Pointers
         allBuildings->insert({daycareBuildings->at(i).getID(), &(daycareBuildings->at(i))});
         densityData->at(daycareBuildings->at(i).getLocation()[0]).at(daycareBuildings->at(i).getLocation()[1]).addBuilding(&(daycareBuildings->at(i)));
@@ -349,13 +349,13 @@ void BuildingGenerator::generateBuildings(double businessSizeProbablities[6], do
         totalBuinessPopulation+=capacity;
     }
     
-    for(int i = 0; i< medicalBuildings->size();i++){
+    for (size_t i = 0; i< medicalBuildings->size();i++) {
         //Set up Pointers
         allBuildings->insert({medicalBuildings->at(i).getID(), &(medicalBuildings->at(i))});
         densityData->at(medicalBuildings->at(i).getLocation()[0]).at(medicalBuildings->at(i).getLocation()[1]).addBuilding(&(medicalBuildings->at(i)));
     }
     
-    for(int i = 0; i< businessBuildings->size();i++){
+    for (size_t i = 0; i< businessBuildings->size();i++){
         //Set up Pointers
         //std::cout<<"Setting up Business Pointers"<<std::endl;
         allBuildings->insert({businessBuildings->at(i).getID(), &(businessBuildings->at(i))});
@@ -368,8 +368,8 @@ void BuildingGenerator::generateBuildings(double businessSizeProbablities[6], do
 
 void BuildingGenerator::assignHomes(Population &pop){
     //Assign Families Locations
-    int x=0;
-    int y=0;
+    size_t x = 0;
+    size_t y = 0;
     int notAssigned=0;
     float oldRatio=0;
     std::cout << "Assigning Homes to Population" << std::endl;
@@ -584,7 +584,7 @@ void BuildingGenerator::importBuildings(std::string fileLocation){
         updateStatistics(type, maxCap);
     }
 
-    for(int i = 0; i< otherBuildings->size();i++){
+    for (size_t i = 0; i < otherBuildings->size(); i++){
         //Set up Pointers
         allBuildings->insert({otherBuildings->at(i).getID(), &(otherBuildings->at(i))});
         densityData->at(otherBuildings->at(i).getLocation()[0]).at(otherBuildings->at(i).getLocation()[1]).addBuilding(&(otherBuildings->at(i)));
@@ -592,7 +592,7 @@ void BuildingGenerator::importBuildings(std::string fileLocation){
             &(otherBuildings->at(i))});
     }
     
-    for(int i = 0; i< daycareBuildings->size();i++){
+    for (size_t i = 0; i < daycareBuildings->size(); i++) {
         //Set up Pointers
         allBuildings->insert({daycareBuildings->at(i).getID(), &(daycareBuildings->at(i))});
         densityData->at(daycareBuildings->at(i).getLocation()[0]).at(daycareBuildings->at(i).getLocation()[1]).addBuilding(&(daycareBuildings->at(i)));
@@ -601,7 +601,7 @@ void BuildingGenerator::importBuildings(std::string fileLocation){
 
     }
     
-    for(int i = 0; i< schoolBuildings->size();i++){
+    for (size_t i = 0; i < schoolBuildings->size(); i++) {
         //Set up Pointers
         allBuildings->insert({schoolBuildings->at(i).getID(), &(schoolBuildings->at(i))});
         densityData->at(schoolBuildings->at(i).getLocation()[0]).at(schoolBuildings->at(i).getLocation()[1]).addBuilding(&(schoolBuildings->at(i)));
@@ -609,7 +609,7 @@ void BuildingGenerator::importBuildings(std::string fileLocation){
             &(schoolBuildings->at(i))});
     }
     
-    for(int i = 0; i< medicalBuildings->size();i++){
+    for (size_t i = 0; i < medicalBuildings->size(); i++) {
         //Set up Pointers
         allBuildings->insert({medicalBuildings->at(i).getID(), &(medicalBuildings->at(i))});
         densityData->at(medicalBuildings->at(i).getLocation()[0]).at(medicalBuildings->at(i).getLocation()[1]).addBuilding(&(medicalBuildings->at(i)));
@@ -617,7 +617,7 @@ void BuildingGenerator::importBuildings(std::string fileLocation){
             &(medicalBuildings->at(i))});
     }
     
-    for(int i = 0; i< businessBuildings->size();i++){
+    for (size_t i = 0; i < businessBuildings->size(); i++) {
         //Set up Pointers
         allBuildings->insert({businessBuildings->at(i).getID(), &(businessBuildings->at(i))});
         densityData->at(businessBuildings->at(i).getLocation()[0]).at(businessBuildings->at(i).getLocation()[1]).addBuilding(&(businessBuildings->at(i)));
